feat(env): add COOKY_USER and COOKY_PASSWORD variables
Added new environment variables `COOKY_USER` and `COOKY_PASSWORD` to the `.env.example` file. These variables will be used to create a new user with specified username and password in the PostgreSQL database for the app.

Also updated the `compose.yml` file to use these new variables when configuring the PostgreSQL service. The script `initdb.sh` has been added, which creates a new user with the provided credentials and grants all privileges on the database.